% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_signals.R
\name{get_SIGRTs}
\alias{get_SIGRTs}
\title{Available SIGRTs}
\usage{
get_SIGRTs()
}
\value{
data.frame
}
\description{
Data.frame containing all available Real-Time UNIX signals:
their number as well as their name as "SIGRTMIN" and "SIGRTMAX"
}
