% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGRT_check.R
\name{SIGRT_check}
\alias{SIGRT_check}
\title{Check if chosen signal have been received}
\usage{
SIGRT_check(signal, refresh = F)
}
\arguments{
\item{signal}{Valid UNIX SIGRT signal to listen for. Either its integer value
or name beginning with "SIGRTMIN" or "SIGRTMAX".}

\item{*boolean*}{indicating if the flag for given signal should be set to \code{0}
after check}
}
\description{
Check if chosen signal have been received
}
