% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGRT_get.R
\name{SIGRT_get}
\alias{SIGRT_get}
\title{Get data about available SIGRTs}
\usage{
SIGRT_get()
}
\value{
data.frame
}
\description{
Data.frame containing all available Real-Time UNIX signals:
their number as well as their name as "SIGRTMIN" and "SIGRTMAX"
}
