% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGRT_send.R
\name{SIGRT_send}
\alias{SIGRT_send}
\title{Send SIGRT signal to process}
\usage{
SIGRT_send(signal, pid = Sys.getpid())
}
\arguments{
\item{signal}{Valid UNIX SIGRT signal to listen for. Either its integer value
or name beginning with "SIGRTMIN" or "SIGRTMAX".}

\item{pid}{\emph{integer} with PID of the receiving process. Defaults to value of
\code{Sys.getpid()}, sending signal to current process.}

\item{retrun}{\emph{boolean} indicating if signal have been sent}
}
\description{
Providing the process PID send one of the valid SIGRT signals
to that process
}
