% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGRT_check.R
\name{SIGRT_check}
\alias{SIGRT_check}
\title{Check if chosen signal have been received}
\usage{
SIGRT_check(signal, refresh = F)
}
\arguments{
\item{signal}{Valid UNIX SIGRT signal to listen for. Either its integer value
or name beginning with "SIGRTMIN" or "SIGRTMAX".}

\item{refresh}{\emph{boolean} indicating if the flag for given signal should be set to \code{0}
after making the check}
}
\description{
Check if chosen signal have been received
}
\examples{
# Begin listening for signal SIGRTMIN+1 and SIGRTMAX-3
SIGRT_listen("SIGRTMIN+1")
SIGRT_listen("SIGRTMAX-3")

# Check if there were no signals received
SIGRT_check("SIGRTMIN+1")
SIGRT_check("SIGRTMAX-3")

# Send some signals to itself
for (i in 1:4)
  SIGRT_send("SIGRTMIN+1")
for (i in 1:2)
  SIGRT_send("SIGRTMAX-3")

# Check if correct number of signals were received
SIGRT_check("SIGRTMIN+1")

# refresh flag for SIGRTMAX-3
SIGRT_check("SIGRTMAX-3", refresh = T)
SIGRT_check("SIGRTMAX-3")

}
