% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigListener.R
\name{SIGRT_listen}
\alias{SIGRT_listen}
\title{Began listening for SIGRT signal}
\usage{
SIGRT_listen(signal)
}
\arguments{
\item{signal}{Valid UNIX SIGRT signal to listen for. Either its integer value
or name beginning with "SIGRTMIN" or "SIGRTMAX".}
}
\value{
boolean indicating if initializing procedure began successfully
}
\description{
Before running checks for particular signal, it is mandatory to first call
this function - it installs the custom handler for specified signal.

As available SIGRT signals vary from system to system, you can check the
list of available signals with \code{\link[=get_SIGRTs]{get_SIGRTs()}}.
}
